fixed a minor issue where null causes table to fail
