+++ src/common/adapters/RedisDataAdapter/scripts.ts
@@ -1,6 +1,6 @@
 import BaseDataScript, { getDivider } from 'src/common/adapters/BaseDataAdapter/scripts';
 import { SqlAction, SqluiCore } from 'typings';
-
+import { getClientOptions } from 'src/common/adapters/RedisDataAdapter/utils';
 export const REDIS_ADAPTER_PREFIX = 'db';
 
 const formatter = 'js';
@@ -296,9 +296,9 @@ export class ConcreteDataScripts extends BaseDataScript {
     query: SqluiCore.ConnectionQuery,
     language: SqluiCore.LanguageMode,
   ) {
-    let connectionString = connection.connection;
-    let sql = query.sql;
-    let database = query.databaseId;
+    const clientOptions = getClientOptions(connection.connection);
+    const sql = query.sql;
+    const database = query.databaseId;
 
     switch (language) {
       case 'javascript':
@@ -308,6 +308,15 @@ const { createClient, RedisClientType } = require('redis');
 
 async function _doWork(){
   try {
+    const db = await new Promise((resolve, reject) => {
+      const client = createClient(${JSON.stringify(clientOptions)});
+      client.connect();
+      client.on('ready', () => resolve(client));
+      client.on('error', (err) => reject(err));
+    });
+
+    const res = await ${sql};
+    console.log(res);
   } catch(err){
     console.log('Failed to connect', err);
   }
